initial commit for re-working data
