+++ SciStreams/analyses/XSAnalysis/DataRQconv.py
@@ -7,289 +7,6 @@ from .tools import roundbydigits
 import numpy as np
 
 
-# Calibration
-################################################################################
-class CalibrationRQconv(Calibration):
-    """
-    The geometric claculations used here are described in Yang, J Synch Rad (2013) 20, 211–218
-    http://dx.doi.org/10.1107/S0909049512048984
-    (See Appendix A)
-
-    For the rotations, let's assume that the detector is initially
-    oriented in the x-y plane, with the xray beam traveling along the
-    negative z direction (as figure 5 in paper).
-
-    The symbols in the paper translate to the following variables:
-        chi : det_phi
-            This can be thought of the orientation of the detector about
-            the beam center (with all other angles zero).
-            A positive motion of this results in a counter-clockwize
-            rotation of the detector in the x-y plane (or a clockwise
-            rotation of the coordinates).
-        phi : det_orient
-            this angle specifies angle of the axis with which the
-            detector tilts. The tilt itself is specified by tau (next
-            variable).
-        tau : det_tilt
-            This is the amount of tilt of the detector about some axis
-            that is in the plane of the detector. The axis has an extra
-            degree of freedom defined by p
-            defined. It appears to be the negative of the axis defined
-            in the paper (TODO : to verify)
-            Note that no tilt (in tau) results in no movement
-                (regardless of phi).
-
-
-    Just to rephrase what is said in paper. When he says that chi (or
-    det_phi) is only for "grazing-incident measurements", he means that
-    the data will not be azimuthally averaged. Note that there are still
-    plenty of SAXS measurements where this tilt is still important.
-
-    """
-    def __init__(self, wavelength_A=None, distance_m=None, pixel_size_um=None,
-                 x0=None, y0=None, width=None, height=None, det_orient=0.,
-                 det_tilt=0., det_phi=0., incident_angle=0., sample_normal=0.):
-
-        self.det_orient = det_orient
-        self.det_tilt = det_tilt
-        self.det_phi = det_phi
-
-        self.incident_angle = incident_angle
-        self.sample_normal = sample_normal
-
-
-        self.rot_matrix = None
-        super().__init__(wavelength_A=wavelength_A, distance_m=distance_m,
-                         pixel_size_um=pixel_size_um, height=height,
-                         width=width, x0=x0, y0=y0)
-
-
-    # Experimental parameters
-    ########################################
-
-    def set_angles(self, det_orient=0, det_tilt=0, det_phi=0, incident_angle=0., sample_normal=0.):
-        self.det_orient = det_orient
-        self.det_tilt = det_tilt
-        self.det_phi = det_phi
-
-        self.incident_angle = incident_angle
-        self.sample_normal = sample_normal
-
-    def get_ratioDw(self):
-        ''' ratio of sample to detector distance to width.'''
-        width_mm = self.width*self.pixel_size_um/1000.
-        return self.distance_m/(width_mm/1000.)
-
-    # Maps
-    ########################################
-
-    @property
-    def q_map(self):
-        '''Returns a 2D map of the q-value associated with each pixel position
-        in the detector image.'''
-
-        if self.q_map_data is None:
-            self.generate_maps()
-
-        return self.q_map_data
-
-    @property
-    def angle_map(self):
-        '''Returns a map of the angle for each pixel (w.r.t. origin).
-        0 degrees is vertical, +90 degrees is right, -90 degrees is left.'''
-
-        if self.angle_map_data is not None:
-            self.generate_maps()
-
-        return self.angle_map_data
-
-    @property
-    def qx_map(self):
-        if self.qx_map_data is None:
-            self.generate_maps()
-
-        return self.qx_map_data
-
-    @property
-    def qy_map(self):
-        if self.qy_map_data is None:
-            self.generate_maps()
-
-        return self.qy_map_data
-
-    @property
-    def qz_map(self):
-        if self.qz_map_data is None:
-            self.generate_maps()
-
-        return self.qz_map_data
-
-    @property
-    def qr_map(self):
-        if self.qr_map_data is None:
-            self.generate_maps()
-
-        return self.qr_map_data
-
-    # r_map already defined in parent object
-
-
-    def generate_maps(self):
-        """
-        calculate all coordinates (pixel position as well as various derived values)
-        all coordinates are stored in 2D arrays, as is the data itself in Data2D
-        """
-        print("Generating qmaps (expensive computation)")
-        self.calc_rot_matrix()
-
-        (w,h) = (self.width, self.height)
-        # y is columnds, x is rows
-        self.Y, self.X = np.meshgrid(np.arange(h), np.arange(w), indexing='ij')
-        X = self.X.ravel()
-        Y = self.Y.ravel()
-
-        (Q, Phi, Qx, Qy, Qz, Qr, Qn, FPol, FSA) = self.calc_from_XY(X, Y, calc_cor_factors=True)
-
-        self.q_map_data = Q.reshape((h, w))
-
-        self.qx_map_data = Qx.reshape((h, w))
-        self.qy_map_data = Qy.reshape((h, w))
-        self.qz_map_data = Qz.reshape((h, w))
-
-        self.qr_map_data = Qr.reshape((h, w))
-        self.qz_map_data = Qn.reshape((h, w))
-        self.angle_map_data = np.degrees( Phi.reshape((h, w)) )
-        self.FPol_map_data = FPol.reshape((h, w))
-        self.FSA_map_data = FSA.reshape((h, w))
-
-
-
-
-    # q calculation
-    ########################################
-
-    def calc_from_XY(self, X, Y, calc_cor_factors=False):
-        """
-        calculate Q values from pixel positions X and Y
-        X and Y are 1D arrays
-        returns reciprocal/angular coordinates, optionally returns
-        always calculates Qr and Qn, therefore incident_angle needs to be set
-        Note that Phi is saved in radians; but the angles in ExpPara are in degrees
-        """
-        if self.rot_matrix is None:
-            raise ValueError('the rotation matrix is not yet set.')
-
-        # the position vectors for each pixel, origin at the postion of beam impact
-        # R.shape should be (3, w*h), but R.T is more convinient for matrix calculation
-        # RT.T[i] is a vector
-        # x0, y0 is considerence the origin, which is also beam center
-        RT = np.vstack((X - self.x0, -(Y - self.y0), 0.*X))
-
-        # get distance from sample in number of pixels (detector coordinates)
-        dr = self.get_ratioDw()*self.width
-        # position vectors in lab coordinates, sample at the origin
-        [X1, Y1, Z1] = np.dot(self.rot_matrix, RT)
-        Z1 -= dr
-
-        # angles
-        r3sq = X1*X1+Y1*Y1+Z1*Z1
-        r3 = np.sqrt(r3sq)
-        r2 = np.sqrt(X1*X1+Y1*Y1)
-        Theta = 0.5*np.arcsin(r2/r3)
-        Phi = np.arctan2(Y1, X1) + np.radians(self.sample_normal)
-
-        Q = 2.0*self.get_k()*np.sin(Theta)
-
-        # lab coordinates
-        Qz = Q*np.sin(Theta)
-        Qy = Q*np.cos(Theta)*np.sin(Phi)
-        Qx = Q*np.cos(Theta)*np.cos(Phi)
-
-        # convert to sample coordinates
-        alpha = np.radians(self.incident_angle)
-        Qn = Qy*np.cos(alpha) + Qz*np.sin(alpha)
-        Qr = np.sqrt(Q*Q-Qn*Qn)*np.sign(Qx)
-
-        if calc_cor_factors==True:
-            FPol = (Y1*Y1+Z1*Z1)/r3sq
-            FSA = np.power(np.fabs(Z1)/r3, 3)
-            return (Q, Phi, Qx, Qy, Qz, Qr, Qn, FPol, FSA)
-        else:
-            return (Q, Phi, Qx, Qy, Qz, Qr, Qn)
-
-
-    # Rotation calculation
-    ########################################
-
-
-    def RotationMatrix(self, axis, angle):
-        if axis=='x' or axis=='X':
-            rot = np.asarray(
-                [[1., 0., 0.],
-                [0., np.cos(angle), -np.sin(angle)],
-                [0., np.sin(angle),  np.cos(angle)]])
-        elif axis=='y' or axis=='Y':
-            rot = np.asarray(
-                [[ np.cos(angle), 0., np.sin(angle)],
-                [0., 1., 0.],
-                [-np.sin(angle), 0., np.cos(angle)]])
-        elif axis=='z' or axis=='Z':
-            rot = np.asarray(
-                [[np.cos(angle), -np.sin(angle), 0.],
-                [np.sin(angle),  np.cos(angle), 0.],
-                [0., 0., 1.]])
-        else:
-            raise ValueError('unknown axis %s' % axis)
-
-        return rot
-
-
-    def calc_rot_matrix(self):
-
-        # First rotate detector about x-ray beam
-        tm1 = self.RotationMatrix('z', np.radians(-self.det_orient))
-
-        # Tilt detector face (so that it is not orthogonal to beam)
-        tm2 = self.RotationMatrix('y', np.radians(self.det_tilt))
-
-        # Rotate detector about x-ray beam
-        tm3 = self.RotationMatrix('z', np.radians(self.det_orient+self.det_phi))
-        self.rot_matrix = np.dot(np.dot(tm3, tm2), tm1)
-
-
-    # End class CalibrationRQconv(Calibration)
-    ########################################
-
-@normalize_token.register(CalibrationRQconv)
-def tokenize_calibrationrqconv(self):
-    '''
-        This function will first inherit the tokenization from Calibration
-        then run it for the RQConv case.
-        It then runs it for the new arguments and appends to the args list.
-
-        It will return something like:
-            'list, [1,1,1], 'list', [1,3,4] etc
-            it looks messy but at least it is hashable
-    '''
-    # inherit dispatch from Calibration object
-    #calib_norm = normalize_token.dispatch(Calibration)
-    #args = calib_norm(self)
-    from .Data import tokenize_calibration
-    args = tokenize_calibration(self)
-    # finally now tokenize the rest
-    newargs = list()
-    # round
-    newargs.append(roundbydigits(self.det_orient, 3))
-    newargs.append(roundbydigits(self.det_tilt, 3))
-    newargs.append(roundbydigits(self.det_phi, 3))
-    newargs.append(roundbydigits(self.incident_angle, 3))
-    newargs.append(roundbydigits(self.sample_normal, 3))
-    newargs.append(roundbydigits(self.rot_matrix, 3))
-    newargs = normalize_token(newargs)
-    args = (args, newargs)
-
-    return args
-
 
 ################ Below is not to be used
 ###   These are just the experimental parameters  #################
